refactor: split file manager session helpers
