+++ src/backend/ssh/file-manager.ts
@@ -26,6 +26,15 @@ import {
   modeToPermissions,
 } from "./file-manager-utils.js";
 import { createJumpHostChain } from "./jump-host-chain.js";
+import {
+  ChannelOpenSerializer,
+  execChannel,
+  execWithSudo,
+  execWithSudoBuffer,
+  getSessionSftp,
+  type PendingTOTPSession,
+  type SSHSession,
+} from "./file-manager-session.js";
 
 const app = express();
 
@@ -42,205 +51,10 @@ app.use((_req, res, next) => {
 const authManager = AuthManager.getInstance();
 app.use(authManager.createAuthMiddleware());
 
-// Serializes SSH channel open requests so only one channel negotiation is
-// in-flight at a time per session. Once the channel is established the slot
-// is released immediately so the next open can proceed; the channels
-// themselves remain open concurrently (one exec per command is short-lived,
-// the SFTP channel is long-lived but only opened once).
-class ChannelOpenSerializer {
-  private tail: Promise<void> = Promise.resolve();
-
-  // Enqueue an action that opens a channel. The action runs after the previous
-  // one completes (success or failure). Returns a promise that resolves with
-  // the action's result.
-  run<T>(action: () => Promise<T>): Promise<T> {
-    const next = this.tail.then(
-      () => action(),
-      () => action(), // run even if the previous open failed
-    );
-    // Advance tail past this slot (swallow result so the chain keeps going)
-    this.tail = next.then(
-      () => {},
-      () => {},
-    );
-    return next;
-  }
-}
-
-interface SSHSession {
-  client: SSHClient;
-  isConnected: boolean;
-  lastActive: number;
-  timeout?: NodeJS.Timeout;
-  activeOperations: number;
-  sudoPassword?: string;
-  sftp?: import("ssh2").SFTPWrapper;
-  sftpPending?: Promise<import("ssh2").SFTPWrapper>;
-  channelOpener: ChannelOpenSerializer;
-  poolKey?: string;
-  userId?: string;
-}
-
-interface PendingTOTPSession {
-  client: SSHClient;
-  finish: (responses: string[]) => void;
-  config: import("ssh2").ConnectConfig;
-  createdAt: number;
-  sessionId: string;
-  hostId?: number;
-  ip?: string;
-  port?: number;
-  username?: string;
-  userId?: string;
-  prompts?: Array<{ prompt: string; echo: boolean }>;
-  totpPromptIndex?: number;
-  resolvedPassword?: string;
-  totpAttempts: number;
-  isWarpgate?: boolean;
-}
-
 const sshSessions: Record<string, SSHSession> = {};
 const pendingTOTPSessions: Record<string, PendingTOTPSession> = {};
 // Keyed by "sessionId:path" to prevent concurrent requests for the same path
 const activeListRequests: Record<string, boolean> = {};
-
-function execWithSudo(
-  session: SSHSession,
-  command: string,
-  sudoPassword: string,
-): Promise<{ stdout: string; stderr: string; code: number }> {
-  return execWithSudoBuffer(session, command, sudoPassword).then((result) => ({
-    stdout: result.stdout.toString("utf8"),
-    stderr: result.stderr,
-    code: result.code,
-  }));
-}
-
-function execWithSudoBuffer(
-  session: SSHSession,
-  command: string,
-  sudoPassword: string,
-): Promise<{ stdout: Buffer; stderr: string; code: number }> {
-  return new Promise((resolve) => {
-    const escapedPassword = sudoPassword.replace(/'/g, "'\"'\"'");
-    const sudoCommand = `echo '${escapedPassword}' | sudo -S ${command} 2>&1`;
-
-    execChannel(session, sudoCommand, (err, stream) => {
-      if (err) {
-        resolve({ stdout: Buffer.alloc(0), stderr: err.message, code: 1 });
-        return;
-      }
-
-      const stdoutChunks: Buffer[] = [];
-      let stderr = "";
-
-      stream.on("data", (chunk: Buffer) => {
-        stdoutChunks.push(chunk);
-      });
-
-      stream.stderr.on("data", (chunk: Buffer) => {
-        stderr += chunk.toString();
-      });
-
-      stream.on("close", (code: number) => {
-        let stdout = Buffer.concat(stdoutChunks);
-        const sudoPromptMatch = stdout
-          .toString("utf8", 0, Math.min(stdout.length, 256))
-          .match(/^\[sudo\] password for .+?:\s*/);
-        if (sudoPromptMatch) {
-          stdout = stdout.subarray(Buffer.byteLength(sudoPromptMatch[0]));
-        }
-        resolve({ stdout, stderr, code: code || 0 });
-      });
-
-      stream.on("error", (streamErr: Error) => {
-        resolve({
-          stdout: Buffer.concat(stdoutChunks),
-          stderr: streamErr.message,
-          code: 1,
-        });
-      });
-    });
-  });
-}
-
-function getSessionSftp(
-  session: SSHSession,
-): Promise<import("ssh2").SFTPWrapper> {
-  if (session.sftp) {
-    return Promise.resolve(session.sftp);
-  }
-
-  // Serialization: if a channel open is already in flight, join it
-  if (session.sftpPending) {
-    return session.sftpPending;
-  }
-
-  const openOnce = (): Promise<import("ssh2").SFTPWrapper> =>
-    session.channelOpener.run(
-      () =>
-        new Promise<import("ssh2").SFTPWrapper>((resolve, reject) => {
-          session.client.sftp((err, sftp) => {
-            if (err) return reject(err);
-            session.sftp = sftp;
-            sftp.on("error", () => {
-              session.sftp = undefined;
-            });
-            sftp.on("close", () => {
-              session.sftp = undefined;
-            });
-            resolve(sftp);
-          });
-        }),
-    );
-
-  session.sftpPending = openOnce()
-    .catch((err: Error) => {
-      const isChannelFailure =
-        err.message.toLowerCase().includes("channel open failure") ||
-        err.message.toLowerCase().includes("open failed");
-      if (isChannelFailure) {
-        // Single retry after 500ms for transient server-side rate limiting
-        return new Promise<import("ssh2").SFTPWrapper>((resolve, reject) =>
-          setTimeout(() => openOnce().then(resolve, reject), 500),
-        );
-      }
-      return Promise.reject(err);
-    })
-    .finally(() => {
-      session.sftpPending = undefined;
-    });
-
-  return session.sftpPending;
-}
-
-// Wraps client.exec through the channel serializer so only one SSH channel
-// negotiation is in-flight at a time. The serializer slot is released as soon
-// as the channel is established (not when it closes), so channels run
-// concurrently once open — we only serialize the *open handshake*.
-function execChannel(
-  session: SSHSession,
-  command: string,
-  callback: (
-    err: Error | undefined,
-    stream: import("ssh2").ClientChannel,
-  ) => void,
-): void {
-  session.channelOpener
-    .run(
-      () =>
-        new Promise<import("ssh2").ClientChannel>((resolve, reject) => {
-          session.client.exec(command, (err, stream) => {
-            if (err) return reject(err);
-            resolve(stream);
-          });
-        }),
-    )
-    .then(
-      (stream) => callback(undefined, stream),
-      (err: Error) => callback(err, undefined as never),
-    );
-}
 
 function cleanupSession(sessionId: string) {
   const session = sshSessions[sessionId];
